animal_class/backend/requirements.txt
fastapi
uvicorn
joblib
pillow
numpy
scikit-learn==1.6.1
python-multipart
torch
torchvision
tqdm
